split routes for each interface
