+++ .env.example
@@ -1,7 +1,7 @@
 # Database
 # See docker-compose.yml for default values
 # Format: postgresql://USER:PASSWORD@HOST:PORT/DATABASE
-DATABASE_URL=changeme
+DATABASE_URL=postgresql://openconnect:hackathon2026@localhost:5432/openconnect
 
 # Redis
 REDIS_URL=redis://localhost:6379
