+++ guilds/voice/api/admin.ts
@@ -1,0 +1,158 @@
+import { Router, Request, Response } from 'express';
+import {
+  requireAuth,
+  requireRole,
+  createSuccessResponse,
+  createErrorResponse,
+  prisma,
+} from '@openconnect/shared';
+
+export const voiceAdminRouter = Router();
+
+// ==========================================
+// ADMIN ENDPOINTS (existing)
+// ==========================================
+
+voiceAdminRouter.get('/active-calls', requireAuth, requireRole('facility_admin', 'agency_admin'), async (req: Request, res: Response) => {
+  try {
+    const { facilityId } = req.query;
+    
+    const activeCalls = await prisma.voiceCall.findMany({
+      where: {
+        status: { in: ['ringing', 'connected'] },
+        ...(facilityId ? { facilityId: String(facilityId) } : {}),
+      },
+      include: {
+        incarceratedPerson: true,
+        familyMember: true,
+      },
+      orderBy: { startedAt: 'desc' },
+    });
+
+    res.json(createSuccessResponse(activeCalls));
+  } catch (error) {
+    console.error('Error fetching active calls:', error);
+    res.status(500).json(createErrorResponse({
+      code: 'INTERNAL_ERROR',
+      message: 'Failed to fetch active calls',
+    }));
+  }
+});
+
+voiceAdminRouter.get('/call-logs', requireAuth, async (req: Request, res: Response) => {
+  try {
+    const { facilityId, startDate, endDate, userId, page = '1', pageSize = '20' } = req.query;
+    
+    const skip = (parseInt(String(page)) - 1) * parseInt(String(pageSize));
+    const take = parseInt(String(pageSize));
+
+    const where: Record<string, unknown> = {};
+    if (facilityId) where.facilityId = String(facilityId);
+    if (userId) {
+      where.OR = [
+        { incarceratedPersonId: String(userId) },
+        { familyMemberId: String(userId) },
+      ];
+    }
+    if (startDate || endDate) {
+      where.startedAt = {
+        ...(startDate ? { gte: new Date(String(startDate)) } : {}),
+        ...(endDate ? { lte: new Date(String(endDate)) } : {}),
+      };
+    }
+
+    const [calls, total] = await Promise.all([
+      prisma.voiceCall.findMany({
+        where,
+        include: {
+          incarceratedPerson: true,
+          familyMember: true,
+        },
+        skip,
+        take,
+        orderBy: { startedAt: 'desc' },
+      }),
+      prisma.voiceCall.count({ where }),
+    ]);
+
+    res.json({
+      success: true,
+      data: calls,
+      pagination: {
+        page: parseInt(String(page)),
+        pageSize: take,
+        total,
+        totalPages: Math.ceil(total / take),
+      },
+    });
+  } catch (error) {
+    console.error('Error fetching call logs:', error);
+    res.status(500).json(createErrorResponse({
+      code: 'INTERNAL_ERROR',
+      message: 'Failed to fetch call logs',
+    }));
+  }
+});
+
+voiceAdminRouter.post('/terminate-call/:callId', requireAuth, requireRole('facility_admin', 'agency_admin'), async (req: Request, res: Response) => {
+  try {
+    const { callId } = req.params;
+    
+    const call = await prisma.voiceCall.update({
+      where: { id: callId },
+      data: {
+        status: 'terminated_by_admin',
+        endedAt: new Date(),
+        endedBy: 'admin',
+        terminatedByAdminId: req.user!.id,
+      },
+    });
+
+    res.json(createSuccessResponse({ success: true, call }));
+  } catch (error) {
+    console.error('Error terminating call:', error);
+    res.status(500).json(createErrorResponse({
+      code: 'INTERNAL_ERROR',
+      message: 'Failed to terminate call',
+    }));
+  }
+});
+
+voiceAdminRouter.get('/stats', requireAuth, requireRole('facility_admin', 'agency_admin'), async (req: Request, res: Response) => {
+  try {
+    const { facilityId, date } = req.query;
+    const targetDate = date ? new Date(String(date)) : new Date();
+    const startOfDay = new Date(targetDate);
+    startOfDay.setHours(0, 0, 0, 0);
+    const endOfDay = new Date(targetDate);
+    endOfDay.setHours(23, 59, 59, 999);
+
+    const where: Record<string, unknown> = {};
+    if (facilityId) where.facilityId = String(facilityId);
+
+    const [activeCalls, todayTotal] = await Promise.all([
+      prisma.voiceCall.count({
+        where: {
+          ...where,
+          status: { in: ['ringing', 'connected'] },
+        },
+      }),
+      prisma.voiceCall.count({
+        where: {
+          ...where,
+          startedAt: { gte: startOfDay, lte: endOfDay },
+        },
+      }),
+    ]);
+
+    res.json(createSuccessResponse({ activeCalls, todayTotal }));
+  } catch (error) {
+    console.error('Error fetching stats:', error);
+    res.status(500).json(createErrorResponse({
+      code: 'INTERNAL_ERROR',
+      message: 'Failed to fetch stats',
+    }));
+  }
+});
+
+export default voiceAdminRouter;
